integrated axios to make req to the backend
