+++ ui/src/services/agents.ts
@@ -1,0 +1,6 @@
+import type { publishAgentType } from "@/types/zod";
+import Api from "./api";
+
+export function PublishAgent(agent: publishAgentType) {
+  return Api.post("/agents", agent);
+}
