+++ ui/src/services/mockData.ts
@@ -1,16 +1,4 @@
-export interface AgentData {
-  id: string;
-  name: string;
-  avatarUrl: string;
-  description: string;
-  strategyType: string;
-  roi: number;
-  roiTimeframe: string;
-  riskScore: number;
-  followers: number;
-  trades: number;
-  verified: boolean;
-}
+import type { AgentData } from "./types";
 // Mock agent data
 export const mockAgents: AgentData[] = [
   {
