+++ ui/src/services/types.ts
@@ -1,0 +1,13 @@
+export interface AgentData {
+  id: string;
+  name: string;
+  avatarUrl: string;
+  description: string;
+  strategyType: string;
+  roi: number;
+  roiTimeframe: string;
+  riskScore: number;
+  followers: number;
+  trades: number;
+  verified: boolean;
+}
